+++ .gitignore
@@ -1,3 +1,4 @@
-Music/
-last_albumart.jpg
-list.txt
+/core/*.pyc
+/Music/
+/last_albumart.jpg
+/list.txt
